Modify rest.js to support callbacks and promises
